Extract TransactionInputChecker class
